Refactors App.js/css into App component directory. Imports React library and App.css
